Add a basic /quitquitquit handler to shut things down.
This endpoint will trigger the torrent client to shut down and we wait
for that in the main() loop on a channel exposed by the client. When
we see that the client has shut down, we cancel the Context used by
the main web serving goroutine so we then exit the ListenAndServe()
call. This lets us shut down gracefully.