Add the basis of template handling for the web server endpoints.
We'll embed a filesystem to capture the templates we want to use in
our handler outputs and make those available. For now, we simply add
the root.tmpl.html template and wire it up to the rootHandler (for
"/"). It is functionally equivalent to the HTML we had in place
previously.