Simplify the models in use.
To match the frontend angular setup, we collapse the datastructures we
consider down to just a single models.Torrent. For now, it contains
hash, uri and name. It's not yet clear how we'll reconstruct the uri
cleanly from a running torrent, so we may need to persist that in some
other way if it becomes a problem.