Separate the internal server code to a dedicated controllers/ file.
As we add functionality to persist and restore active torrents, it
makes sense to split out the server functionality into a dedicated
file. We add methods to write and individual torrent file's metainfo
to disk and a second one to read all persisted metainfo files back
into the client. A helper method for accessing the data directory from
the config is added too.