Cache the /torrent component between routings
This preserves the sort order (and later additional component
settings) between different routes. Because routing is done via the
mat-sidenav which lives in the app component, we had to inject the
ActivatedRoute there so it was available to the caching strategy which
we want to have visibility to every routing transition.