Add basic ability for the server to track more metadata
The server now maintains a map of torrent hash to additional
metadata. The metadata includes the original magnet link and whether
or not we think the torrent is actively downloading.

This fucntionality is combined with the call to persist our state to
disk. We still only export the metadata exposed by the torrent
library, but with this additional metadata, we can transition to
storing both our state and the existing state together.