Pass "Done" state from server to client.
We synthesize this based on bytes down being the same as total
bytes. This has a side effect for torrents where we don't yet have the
info. They'll be in "Done" state too, which means they won't get play
or pause buttons in the UI. This is good as we can't start them until
the initial info is gathered anyway.