Better handling for metadata file cleanup on dropTorrent()
If we can't remove the file, it's likely because it isn't there. And
that happens when either something has changed outside of the process
or there is still a goroutine blocked on <-t.GotInfo(). In the latter
case, the bookkeeping in the process will get wonky if we remove the
metadata state from the server so we leave it there. If the GotInfo
call returns, that goroutine will complete the rest of the metadata
setup. If it hangs, a server restart will get things back in
shape. This is the least bad approach to that because we don't have a
good way to handle hung torrent lookups.