Ensure that torrent control buttons always stack right
This ensure the delete option is always aligned. Do this by adding a
display flex styling to the div that holds the buttons and a span with
auto flex prior to adding any buttons. The span will then fill the
space and push the buttons to the right which aligns them how we want
them to be.