For play/pause events, update the table data dynamically.
When the controls component emits an event that is PLAY or
PAUSE (!DELETE), filter the list of torrents to remove (by id) the
original version of the updated torrent and then re-add with the new
data returned with the event. This lets the table display update
quickly without waiting for the next auto-refresh or forcing a full
reload of the page.