Start to improve backend init and shutdown mechanics.
Let's start to make shutdown a bit cleaner as we start to begin
building in handling for signals from the torrent library when
downloads complete. We'll need a goroutine that also gets shutdown. We
also move the publich Shutdown handler from controllers.go to
server.go because it's anot a handler (and won't be).

At the same time, move some of the server init into a dedicated
function. From that, we'll fork a long running goroutine to handle
signals from torrent completion, etc.