Allow sorting by progress on both simple and detail view.
Adding a custom sortingDataAccess method to the material table data
source allows us to sort by percent completion on the progress
column.

We may want to add a second level sort when this is in use
because equivalent percentages don't see a stable sort on name which
can appear janky in the UI. For now, this is handy enough to submit
and iterate.