Allow configuring whether torrent storage paths include torrent id.
Because we'll need to be able to build a full path to where torrent
data lives, we need to have explicit understanding of whether this
option is set or not because it influences the final path that we'll
construct. Alternately, we could teach the torrent.Session object to
give us this info.